Use implicit cast in future_impl
Type promotion does not work fully to promote FutureOr<T> to T after
testing for Future<T>. There are a couple of unnecessary casts `as T`
because of this.

This change uses `as dynamic` instead of `as T` so that the check is
implicit. On dart2js under `-O3` implicit checks are omitted. Using
`as dynamic` ensures that the undetected error in user code does not
cause an unguarded error in the runtime.

Change-Id: I50f6855d05dce1cca202528d4d34d9f5d6e96ec6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/143328
Reviewed-by: Leaf Petersen <leafp@google.com>
Commit-Queue: Stephen Adams <sra@google.com>