[nnbd_migration] allow edits to be done incrementally.
Mutate datastructures to match disk contents. Leverages offset mapper,
by rebasing the migration mapping onto the disk changes which occur in
sequence.

Add an integration test for preview site tests that confirms multiple
changes workflow works, including regions and navigation, including
using the changed edit offsets.

One thing that is not updated is traces. It seems that trace entries are
misusing "offset" by storing column values. Without that, adjusting
trace entries is not trivial. I have not looked into how difficult it
would be to correct this. Traces still *work*, they just may point to an
incorrect column. This is not easy to notice, but worth fixing in the
next CLs.

Another potential consistency improvement would be to use the offset
mapper to handle regions so that they do not need to be updated this
way. This would be more consistent with how other things like navigation
did not require updates once the offset mapper was done correctly.

Another quirk is that this API for performing edits does not *necessarily*
lend itself to reruns. For this CL, I decided to check when the edit is
"insertion only" and therefore easy to do incrementally. However, the API
allows for things like mass deletions that might be trickier to handle in
general. In this case, it will still do a rerun.

Change-Id: Ib88bb32592f54502d69b58ccab71ffb0b5f8040e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/143325
Commit-Queue: Mike Fairhurst <mfairhurst@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>