Add a static error test for default catch type
Static error test for https://github.com/dart-lang/language/issues/922

If the type of the error argument was `dynamic` there would be no error
for calling a method that does not exist.

If the type of the error argument was `Object?` there would be an error
for the call to `_takesObject`.

Change-Id: I3cec1a8ae2903887e5da086cdeecd4008ef5d667
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/143329
Commit-Queue: Nate Bosch <nbosch@google.com>
Reviewed-by: Leaf Petersen <leafp@google.com>