[vm/compiler] Fix non-convergence in constant propagator.
Given cyclic phi which participates in strict or equality comparison
with its own argument

  y <- phi(x, y);
  z <- y == x

CP pass would sometimes fail to converge because it would mark the phi
as unwrapped and then add it to the worklist, which would cause the
phi uses to be revisited: equality would mark the phi again and then
phi will be visited as its own use, which would add it to the worklist
because it is marked and so on.

The fix is to only add marked phis to worklist when reachability of
one of the predecessor blocks changes and that affects whether phi
can be unwrapped or not.

Test is included in a separate CL to make this fix easy to cherry pick.

Fixes https://github.com/flutter/flutter/issues/53903

Change-Id: I9e09a155d7109b7e19ba3ae5aff3fab57dc67fb8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/143581
Reviewed-by: Martin Kustermann <kustermann@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>