[vm] Drop virtual addresses from non-symbolic stack frames.
Virtual addresses are only guaranteed to be valid for the original
shared library that was loaded as a snapshot. This can lead to confusion
if someone attempts to use the virtual address with separately saved
debugging information. This information may not include all the same
sections as the snapshot or include them in the same order and thus may
have different virtual addresses for the instructions sections.

In fact, this happens right now: we don't include a .bss section in the
debugging information, as there's no code in the text sections that
needs relocation. Since the .bss section comes before the .text
sections, this means the (empty) .text sections in the debugging
information have different virtual offsets than the .text sections in
the snapshot.

At this point, we print enough information in non-symbolic stack traces
that the virtual address can be calculated from the absolute address if
truly needed, whether the snapshot was loaded by the OS or by our
non-native ELF loader, so stop printing it to avoid this confusion.

Change-Id: I7949524cc3bb2f510d6812d226b66806939d0da7
Cq-Include-Trybots: luci.dart.try:vm-kernel-precomp-linux-release-x64-try,vm-kernel-precomp-linux-product-x64-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/143801
Commit-Queue: Tess Strickland <sstrickl@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>
Reviewed-by: Daco Harkes <dacoharkes@google.com>