Migration: apply guards to conditional branch expression values.
Consider the following code:

  int f(int i) => i == null ? null : i + 1;

Previously, migration would mark the return type of `f` as nullable
even if `i` was non-nullable.  This happened because the edge
connecting the explicit `null` (in the "then branch" of the
conditional) to the return type of `f` is created after visiting the
conditional expression, so the condition guards are no longer in
scope.

With this CL, we create an extra graph node when needed to ensure that
the guard is applied to the expression type.

Change-Id: Ie00eea0b431905241612c6661d69bdb55c0910d2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/144121
Reviewed-by: Mike Fairhurst <mfairhurst@google.com>