[vm/compiler] Fix linker error when building with -mdebug -d0
For some reason constexpr seems to need a definition, because the linker
complains about not being able to find the symbol:

    % tools/gn.py -ax64 -mdebug -d0
    % tools/build.py -ax64 -mdebug
    ...
    ld.lld: error: undefined symbol: dart::FlowGraphBuilderHelper::kPhiSelfReference
    >>> referenced by constant_propagator_test.cc:53 (../../runtime/vm/compiler/backend/constant_propagator_test.cc:53)
    ...

We may also want to consider passing `-fmerge-all-constants` explicitly
in the future - though it is not standards compliant.

See also http://crbug.com/829795.

Change-Id: I1cc125a01751c8a83092b033e80afdcc591c7b92
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/144287
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>