Add dart::ComponentContext() for Fuchsia
This allows the program to get at the singleton component context
for the currently running component.  The component context can
be meaningfully instantiated only once; and since dart doesn't
have a predefined way to wire platform-specific values into
platform-specific code, we instantiate a singleton here, so all
code that is interested can use it.

The users of this functionality will be the dart and flutter
fuchsia runners.  See issue
https://github.com/dart-lang/sdk/issues/41523 for details.

Tested:
  manual compile for fuchsia, check behavior.

Change-Id: Ib4e15af376ecb932c2d2b28d6f099383d53ccc84
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/136524
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>