[ddc] Remove type and null checks in dart:_runtime
Only removing checks for common calls within the type system that
we know are always safe.

Added `@notNull` on arguments typed to be non-nullable (in the nnbd
version) to avoid producing the null checks.

Added `Object` type to arguments to match the return type and avoid
casts when returning.

Change-Id: I05de45a8d923497f7340348ef2c0b9f72da5414b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/143862
Reviewed-by: Mark Zhou <markzipan@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>