Migration: add the ability for the preview server to be turned off.
This is necessary for unit testing, and probably good practice for
normal usage; we don't want to fire up a server even briefly if the
user specifies `--no-web-preview` when running the migration tool.

Not yet wired up to dartfix--probably I won't bother to do this since
migration is being decoupled from dartfix and a new CLI is being
implemented for it.

Change-Id: I79862d9baf328d0f428f63955fb5918c94186ccb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/144860
Reviewed-by: Janice Collins <jcollins@google.com>