UTF-8 decoder using a state machine.
Two-pass decoder: the first pass scans through the input to compute
the length of the resulting string and which decoder to use, and the
second pass does the actual decoding.

The same decoder is used for both one-shot and chunked decoding, and
both with and without allowMalformed. If there is an error in the input
and allowMalformed is true, it starts over with a general decoder that
supports malformed input and allocates space as it goes along.

JS targets go directly to the general decoder, as the two-pass approach
is not beneficial here.

Three pieces of the decoder are designed to be pluggable by patches to
optimize the performance further:
- scan, running the first pass of the conversion.
- decode8, decoding Latin1 data into a OneByteString.
- decode16, decoding arbitrary data into a TwoByteString.

Improves decoding speed, especially for complex input (many multi-byte
characters). Observed speed increases are approximately:
 - dart2js: up to 40%
 - VM JIT:  up to 260%
 - VM AOT:  up to 130%

The constant overhead of calling the UTF-8 decoder is also significantly
reduced for dart2js.

Code size for dart2js is slightly reduced compared to the old decoder.

ASCII inputs currently see a slight speed decrease for VM targets, which
will be fixed in https://dart-review.googlesource.com/c/sdk/+/145460

This is part of the implementation of the breaking change described at
https://github.com/dart-lang/sdk/issues/41100

Closes https://github.com/dart-lang/sdk/issues/28832
Closes https://github.com/dart-lang/sdk/issues/31954

Ideas for further improvements to the decoder are collected in
https://github.com/dart-lang/sdk/issues/41734

Change-Id: I3c5bb84e8d6783231680a9d34d6c38e8a28ab112
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/142025
Reviewed-by: Stephen Adams <sra@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>