Migration: remove unnecessary TODO comment about supporting `--debug` in new CLI.
In the dartfix-based implementation, the only significant effect of
supplying `--debug` was to enable the analysis server's
instrumentation log.  Since we don't have an analysis server layer in
the new CLI, there's no need for this command line option right now.

If we decide we need a `--debug` flag later we can always add it.

Change-Id: Ib729c5793044456ef460cfa7f823c48f5af9e898
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/146445
Reviewed-by: Mike Fairhurst <mfairhurst@google.com>