[release] Allow promotion from one channel to another
The Travis `beta` channel can break when a new stable release is
published and packages use that release as minimum SDK constraint.
Therefore, we want to publish stable releases to beta if the current
beta is less than the current stable release. This change adds a
`--source-channel` option to promote.py that allows publishing a release
across channels.

Example (will promote the `2.8.1` stable release to beta):
```
$ python promote.py promote \
  --dry --channel=beta --source-channel=stable --version=2.8.1
```

Change-Id: Ieac4c33dfe17d4b38060e10ff3aff44494aae40a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/146787
Reviewed-by: William Hesse <whesse@google.com>