Migration: reorganize code in source_edit_diff_formatter.dart
Based on suggestions from
https://dart-review.googlesource.com/c/sdk/+/146442.  Specifically,
the `_Output` class hierarchy has been moved to a new file so that the
classes in it are together, and the remaining classes (in the
DiffStyle hierarchy) are together in source_edit_diff_formatter.dart.
This required making a bunch of things public, so I split `DiffStyle`
into an interface and and impl class to make it easier to see which
method is intended to be used (DiffStyle.formatDiff).  Also, moved the
logic that was previously in _Output._processEdits to
DiffStyleImpl.formatDiff.

Change-Id: Iad448f16233e6752e7b20d1e0cee4eb5c033f498
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/146821
Reviewed-by: Mike Fairhurst <mfairhurst@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>